// src/interfaces/jobApplication.ts
export interface JobApplication {
    firstName: string;
    middleName?: string;
    age: number;
    lastName: string;
    email: string;
    phone: string;
    position: string;
    institution: string;
    degree: string;
    fieldOfStudy: string;
    startDate?: string;
    score: number;
    yearsOfExperience: number;
    status: boolean,
}